verified program

+++ BackOffice/boleiasModule/__init__.py
@@ -9,21 +9,20 @@ rideModule = Blueprint('rideModule', __name__)
 
 @rideModule.route('/ridePage', methods=['GET', 'POST'])
 def toRidePage():
-    if current_user.is_authenticated:
-        print("Noice")
-    else:
-        print("Not Noice")     
     if request.method == 'GET':
         get_Ride= db.session.query(Ride).all()
         return render_template("boleias.html", title="Login", get_Ride = get_Ride)
     elif request.method == 'POST':
         id = request.form.get("id")
-        print(id)
+        
         specify_Ride = db.session.query(Ride).filter(Ride.id==id).first()
         specify_Resert = db.session.query(Reservation).filter(specify_Ride.id==Reservation.ride_id).first()
         
         db.session.delete(specify_Ride)
-        db.session.delete(specify_Resert)
+
+        if(specify_Resert != None):
+            db.session.delete(specify_Resert)
+
         db.session.commit()
     return  jsonify(message="Apagado",status=201)
